requirements.txt
PyYAML>=6.0
numpy>=1.26
pandas
joblib